restructure AIBL utils module
