+++ clinica/iotools/converters/aibl_to_bids/utils/__init__.py
@@ -1,0 +1,16 @@
+"""Utils to convert AIBL dataset in BIDS."""
+
+from .bids import Modality, paths_to_bids
+from .clinical import (
+    create_participants_tsv_file,
+    create_scans_tsv_file,
+    create_sessions_tsv_file,
+)
+
+__all__ = [
+    "create_participants_tsv_file",
+    "create_scans_tsv_file",
+    "create_sessions_tsv_file",
+    "Modality",
+    "paths_to_bids",
+]
